Move incoming networking into separate file
