+++ packages/app/src/game/sagas/networking/saga.ts
@@ -1,186 +1,11 @@
-import io = require("socket.io-client");
-import { eventChannel } from "redux-saga";
-import { call, takeEvery, put, take, fork, all, select } from "@redux-saga/core/effects";
-import { playerListUpdated } from "../../features/playerList/playerListActions";
-import { log } from "../../../log";
-import { AppState } from "../../../store/state";
-import { signIn } from "../../../auth/auth0";
+import { takeEvery, put, fork, all } from "@redux-saga/core/effects";
 import {
-    BoardCommands,
-    PlayerActions, PlayerInfoCommands, BenchCommands,
-    BATTLE_FINISH_EVENT, startBattle,
-
-    validateNickname,
-    IncomingPacketRegistry, OutgoingPacketRegistry, ConnectionStatus,
-    ServerToClientPacketDefinitions, ServerToClientPacketOpcodes, ServerToClientPacketAcknowledgements, AuthenticateResponse,
-    ClientToServerPacketDefinitions, ClientToServerPacketOpcodes, ClientToServerPacketAcknowledgements, SEND_PLAYER_ACTIONS_PACKET_RETRY_TIME_MS,
-    ServerToClientLobbyPacketDefinitions, ServerToClientLobbyPacketOpcodes, ServerToClientLobbyPacketAcknowledgements, GameEvents
-} from "@creature-chess/shared";
-import { AuthSelectors } from "../../../auth";
-import {
-    clearAnnouncement, closeOverlay, FindGameAction, FIND_GAME, openOverlay,
-    updateConnectionStatus, finishGameAction, playersResurrected
-} from "../../../ui/actions";
-import { joinLobbyAction, updateLobbyPlayerAction } from "../../../lobby/store/actions";
-import { GamePhase } from "@creature-chess/models";
-import { Overlay } from "../../../ui/overlay";
-import { clearSelectedPiece } from "../../features/board/actions";
-
-type Socket = SocketIOClient.Socket;
+    PlayerActions, BATTLE_FINISH_EVENT, OutgoingPacketRegistry, ConnectionStatus,
+    ClientToServerPacketDefinitions, ClientToServerPacketOpcodes, ClientToServerPacketAcknowledgements, SEND_PLAYER_ACTIONS_PACKET_RETRY_TIME_MS} from "@creature-chess/shared";
+import { updateConnectionStatus} from "../../../ui/actions";
+import { incomingGameNetworking } from "./incoming";
 
 type ClientToServerPacketRegsitry = OutgoingPacketRegistry<ClientToServerPacketDefinitions, ClientToServerPacketAcknowledgements>;
-type ServerToClientPacketRegistry = IncomingPacketRegistry<ServerToClientPacketDefinitions, ServerToClientPacketAcknowledgements>;
-
-const subscribe = (
-    registry: ServerToClientPacketRegistry,
-    socket: Socket
-) => {
-    return eventChannel(emit => {
-        socket.on("reconnect_failed", () => {
-            emit(clearAnnouncement());
-            emit(updateConnectionStatus(ConnectionStatus.DISCONNECTED));
-        });
-        socket.on("reconnect_error", () => {
-            emit(clearAnnouncement());
-            emit(updateConnectionStatus(ConnectionStatus.DISCONNECTED));
-        });
-
-        registry.on(
-            ServerToClientPacketOpcodes.PLAYER_LIST_UPDATE,
-            (packet) => {
-                log("[PLAYER_LIST_UPDATE]", packet);
-                emit(playerListUpdated(packet));
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.CARDS_UPDATE,
-            (packet) => {
-                log("[CARDS_UPDATE]", packet);
-                emit(PlayerInfoCommands.updateCardsCommand(packet));
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.MONEY_UPDATE,
-            (packet) => {
-                log("[MONEY_UPDATE]", packet);
-                emit(PlayerInfoCommands.updateMoneyCommand(packet));
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.PHASE_UPDATE,
-            (packet) => {
-                log("[PHASE_UPDATE]", packet);
-
-                emit(GameEvents.gamePhaseStartedEvent(packet.phase, packet.startedAtSeconds));
-
-                switch (packet.phase) {
-                    case GamePhase.PREPARING: {
-                        const { cards, pieces: { board, bench } } = packet.payload;
-
-                        emit(BoardCommands.initialiseBoard(board.pieces));
-                        emit(BenchCommands.initialiseBenchCommand(bench));
-                        emit(PlayerInfoCommands.updateCardsCommand(cards));
-                        emit(PlayerInfoCommands.clearOpponentCommand());
-                        emit(BoardCommands.unlockBoard());
-                        emit(openOverlay(Overlay.SHOP));
-                        emit(clearAnnouncement());
-                        return;
-                    }
-                    case GamePhase.READY: {
-                        const { board, bench, opponentId } = packet.payload;
-
-                        if (board) {
-                            emit(BoardCommands.initialiseBoard(board.pieces));
-                        }
-
-                        emit(BenchCommands.initialiseBenchCommand(bench));
-                        emit(BoardCommands.lockBoard());
-                        emit(closeOverlay());
-                        emit(PlayerInfoCommands.updateOpponentCommand(opponentId));
-                        emit(clearSelectedPiece());
-                        return;
-                    }
-                    case GamePhase.PLAYING: {
-                        emit(startBattle());
-                        return;
-                    }
-                    default:
-                        return;
-                }
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.LEVEL_UPDATE,
-            (packet) => {
-                log("[LEVEL_UPDATE]", packet);
-
-                emit(PlayerInfoCommands.updateLevelCommand(packet.level, packet.xp));
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.JOIN_GAME,
-            ({ fullState }) => {
-                const { money, cards, players, level: { level, xp }, board, bench, phase } = fullState;
-
-                emit(PlayerInfoCommands.updateMoneyCommand(money));
-                emit(PlayerInfoCommands.updateCardsCommand(cards));
-                emit(playerListUpdated(players));
-                emit(PlayerInfoCommands.updateLevelCommand(level, xp));
-                emit(BoardCommands.initialiseBoard(board));
-                emit(BenchCommands.initialiseBenchCommand(bench));
-
-                if (phase) {
-                    emit(GameEvents.gamePhaseStartedEvent(phase.phase, phase.startedAtSeconds));
-                } else {
-                    emit(updateConnectionStatus(ConnectionStatus.RECONNECTED));
-                }
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.FINISH_GAME,
-            (packet) => {
-                log("[FINISH_GAME]", packet);
-
-                emit(finishGameAction(packet.winnerName));
-
-                socket.close();
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.SHOP_LOCK_UPDATE,
-            (packet) => {
-                log("[SHOP_LOCK_UPDATE]", packet);
-
-                emit(PlayerInfoCommands.updateShopLockCommand(packet.locked));
-            }
-        );
-
-        registry.on(
-            ServerToClientPacketOpcodes.PLAYERS_RESURRECTED,
-            ({ playerIds }) => {
-                emit(playersResurrected(playerIds));
-            }
-        );
-
-        // tslint:disable-next-line:no-empty
-        return () => { };
-    });
-};
-
-const readPacketsToActions = function*(incomingRegistry: ServerToClientPacketRegistry, socket: Socket) {
-    const channel = yield call(subscribe, incomingRegistry, socket);
-
-    yield takeEvery(channel, function*(action) {
-        yield put(action);
-    });
-};
 
 const sendPlayerActions = function*(registry: ClientToServerPacketRegsitry) {
     let transmissionInProgress = false;
@@ -267,43 +92,15 @@ const writeActionsToPackets = function*(registry: ClientToServerPacketRegsitry) 
     ]);
 };
 
-export const networking = function*() {
-    // const action: FindGameAction = yield take(FIND_GAME);
+export const networking = function*(socket: SocketIOClient.Socket) {
+    console.log("game networking started");
 
-    // const state: AppState = yield select();
+    const outgoingRegistry = new OutgoingPacketRegistry<ClientToServerPacketDefinitions, ClientToServerPacketAcknowledgements>(
+        (opcode, payload, ack) => socket.emit(opcode, payload, ack)
+    );
 
-    // // this should never happen, but it doesn't hurt to be safe
-    // if (!AuthSelectors.isLoggedIn(state)) {
-    //     signIn();
+    yield put(updateConnectionStatus(ConnectionStatus.CONNECTED));
 
-    //     return;
-    // }
-
-    // const idToken = AuthSelectors.getIdToken(state);
-
-    // let socket: Socket = null;
-
-    // while (socket === null) {
-    //     try {
-    //         socket = yield call(getSocket, action.payload.serverIP, idToken);
-    //     } catch (e) {
-    //         // todo improve this, currently it just refreshes if there's an error.
-    //         signIn();
-
-    //         return;
-    //     }
-    // }
-
-    // const outgoingRegistry = new OutgoingPacketRegistry<ClientToServerPacketDefinitions, ClientToServerPacketAcknowledgements>(
-    //     (opcode, payload, ack) => socket.emit(opcode, payload, ack)
-    // );
-
-    // const incomingRegistry = new IncomingPacketRegistry<ServerToClientPacketDefinitions, ServerToClientPacketAcknowledgements>(
-    //     (opcode, handler) => socket.on(opcode, handler)
-    // );
-
-    // yield put(updateConnectionStatus(ConnectionStatus.CONNECTED));
-
-    // yield fork(readPacketsToActions, incomingRegistry, socket);
-    // yield fork(writeActionsToPackets, outgoingRegistry);
+    yield fork(writeActionsToPackets, outgoingRegistry);
+    yield fork(incomingGameNetworking, socket);
 };
